refactor(svm-sdk): extract BaseSvmSigner base class shared by SvmSigner
Extract an abstract BaseSvmSigner holding the shared send/confirm pipeline and
signer logic; SvmSigner becomes a thin subclass with two override points
(signMessage, skipPreflight). Move the shared transaction types into types.ts.
Behavior-preserving: SvmSigner still defers skipPreflight to the per-tx value.
