+++ typescript/svm-sdk/src/types.ts
@@ -12,6 +12,7 @@ import type { DeployedIsmAddress } from '@hyperlane-xyz/provider-sdk/ism';
 import { isNullish } from '@hyperlane-xyz/utils';
 
 import type { IgpFeeConfig } from './codecs/igp.js';
+import type { InstructionAccountMeta } from './instructions/utils.js';
 
 export type SvmInstruction = Instruction;
 
@@ -73,6 +74,33 @@ export type WithExtraSigners<T> = T & {
 export type AnnotatedSvmTransaction = SvmTransaction & {
   annotation?: string;
 };
+
+/** Transaction input for the submission pipeline — feePayer is provided by the signer. */
+export type SendableSvmTransaction = Omit<SvmTransaction, 'feePayer'>;
+
+/** Shape returned by `buildPrintableTransaction`. */
+export interface PrintableSvmTransaction {
+  annotation?: string;
+  instructions: PrintableSvmInstruction[];
+  computeUnits?: number;
+  transaction_base58: string;
+  message_base58: string;
+  /**
+   * Carried over from `SvmTransaction.waitForSlotAdvance`. A live signer
+   * enforces this itself, but exported (file/Squads) transactions are signed
+   * and submitted by an external executor: when true, that executor MUST wait
+   * for the cluster slot to advance past this transaction's confirmation slot
+   * before submitting the next transaction, otherwise the loader rejects an
+   * extend→upgrade→config sequence sharing a slot.
+   */
+  waitForSlotAdvance?: boolean;
+}
+
+export interface PrintableSvmInstruction {
+  programAddress: Address;
+  accounts?: readonly InstructionAccountMeta[];
+  data?: string;
+}
 
 /**
  * Specifies how to obtain a deployed program address:
